# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A production-ready REST API for restaurant menu management",
        "title": "Restaurant Menu API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.restaurant-menu-api.com/support",
            "email": "support@restaurant-menu-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/restaurants/hours": {
            "get": {
                "description": "Get the restaurant's operating hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get operating hours",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.OperatingHour"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/restaurants/info": {
            "get": {
                "description": "Get the restaurant's basic information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get restaurant information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.RestaurantInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the restaurant's basic information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update restaurant information",
                "parameters": [
                    {
                        "description": "Restaurant data",
                        "name": "restaurant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_handlers.UpdateRestaurantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.RestaurantInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create the restaurant's basic information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create restaurant information",
                "parameters": [
                    {
                        "description": "Restaurant data",
                        "name": "restaurant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_handlers.CreateRestaurantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.RestaurantInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/restaurants/{id}": {
            "delete": {
                "description": "Delete a restaurant by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete a restaurant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Restaurant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_interfaces_handlers.CreateRestaurantRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "address": {
                    "type": "object",
                    "additionalProperties": true
                },
                "contact_info": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "settings": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "internal_interfaces_handlers.UpdateRestaurantRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "type": "object",
                    "additionalProperties": true
                },
                "contact_info": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "settings": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "restaurant-menu-api_internal_domain_entities.Address": {
            "type": "object",
            "additionalProperties": true
        },
...
        "restaurant-menu-api_internal_domain_entities.ContactInfo": {
            "type": "object",
            "additionalProperties": true
        },
...
        "restaurant-menu-api_internal_domain_entities.OperatingHour": {
            "type": "object",
            "properties": {
                "close_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer",
                    "maximum": 6,
                    "minimum": 0
                },
                "id": {
                    "type": "integer"
                },
                "is_closed": {
                    "type": "boolean"
                },
                "open_time": {
                    "type": "string"
                },
                "restaurant_info_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "restaurant-menu-api_internal_domain_entities.Pagination": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "restaurant-menu-api_internal_domain_entities.RestaurantInfo": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.Address"
                },
                "contact_info": {
                    "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.ContactInfo"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "operating_hours": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.OperatingHour"
                    }
                },
                "settings": {
                    "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.Settings"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "restaurant-menu-api_internal_domain_entities.Settings": {
            "type": "object",
            "additionalProperties": true
        },
...
        "restaurant-menu-api_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "restaurant-menu-api_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/restaurant-menu-api_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "restaurant-menu-api_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.Pagination"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}